+++ mayan/apps/storage/tests/mixins.py
@@ -1,8 +1,13 @@
+import glob
 import importlib
 import logging
+import os
 from pathlib import Path
 import shutil
 
+import psutil
+
+from django.conf import settings
 from django.core.files.base import ContentFile
 
 from mayan.apps.common.tests.literals import (
@@ -16,6 +21,7 @@ from mayan.apps.smart_settings.settings import setting_cluster
 from ..classes import DefinedStorage
 from ..compressed_files import Archive
 from ..models import DownloadFile, SharedUploadedFile
+from ..settings import setting_temporary_directory
 from ..utils import mkdtemp
 
 from .literals import (
@@ -70,6 +76,43 @@ class ArchiveClassTestCaseMixin:
             self.assertEqual(
                 file_object.read(), self.member_contents
             )
+
+
+class DescriptorLeakCheckTestCaseMixin:
+    _skip_file_descriptor_test = False
+
+    def _get_process_descriptor_count(self):
+        process = psutil.Process()
+        return process.num_fds()
+
+    def _get_process_descriptors(self):
+        process = psutil.Process()._proc
+        return os.listdir(
+            path='{}/{}/fd'.format(process._procfs_path, process.pid)
+        )
+
+    def setUp(self):
+        super().setUp()
+        self._process_descriptor_count = self._get_process_descriptor_count()
+        self._process_descriptors = self._get_process_descriptors()
+
+    def tearDown(self):
+        if not self._skip_file_descriptor_test:
+            if self._get_process_descriptor_count() > self._process_descriptor_count:
+                raise ValueError(
+                    'File descriptor leak. The number of file descriptors '
+                    'at the end are higher than at the start of the test.'
+                )
+
+            for descriptor in self._get_process_descriptors():
+                if descriptor not in self._process_descriptors:
+                    raise ValueError(
+                        'File descriptor leak. A descriptor was found at '
+                        'the end of the test that was not present at the '
+                        'start of the test.'
+                    )
+
+        super().tearDown()
 
 
 class DownloadFileAPIViewTestMixin:
@@ -141,6 +184,29 @@ class DownloadFileViewTestMixin:
 
     def _request_test_download_file_list_view(self):
         return self.get(viewname='storage:download_file_list')
+
+
+class OpenFileCheckTestCaseMixin:
+    _skip_open_file_leak_test = False
+
+    def _get_open_files(self):
+        process = psutil.Process()
+        return process.open_files()
+
+    def setUp(self):
+        super().setUp()
+
+        self._open_files = self._get_open_files()
+
+    def tearDown(self):
+        if not self._skip_open_file_leak_test:
+            for new_open_file in self._get_open_files():
+                if new_open_file not in self._open_files:
+                    raise ValueError(
+                        'File left open: {}'.format(new_open_file)
+                    )
+
+        super().tearDown()
 
 
 class StorageProcessorTestMixin:
@@ -216,3 +282,46 @@ class StorageSettingTestMixin:
         importlib.reload(storage_module)
 
         return assertion
+
+
+class TempfileCheckTestCasekMixin:
+    # Ignore the jvmstat instrumentation and GitLab's CI .config files.
+    # Ignore LibreOffice fontconfig cache dir.
+    ignore_globs = ('hsperfdata_*', '.config', '.cache')
+
+    def _get_temporary_entries(self):
+        ignored_result = []
+
+        # Expand globs by joining the temporary directory and then flattening
+        # the list of lists into a single list.
+        for item in self.ignore_globs:
+            ignored_result.extend(
+                glob.glob(
+                    os.path.join(setting_temporary_directory.value, item)
+                )
+            )
+
+        # Remove the path and leave only the expanded filename.
+        ignored_result = map(lambda x: os.path.split(x)[-1], ignored_result)
+
+        return set(
+            os.listdir(setting_temporary_directory.value)
+        ) - set(ignored_result)
+
+    def setUp(self):
+        super().setUp()
+        if getattr(settings, 'COMMON_TEST_TEMP_FILES', False):
+            self._temporary_items = self._get_temporary_entries()
+
+    def tearDown(self):
+        if getattr(settings, 'COMMON_TEST_TEMP_FILES', False):
+            final_temporary_items = self._get_temporary_entries()
+            self.assertEqual(
+                self._temporary_items, final_temporary_items,
+                msg='Orphan temporary file. The number of temporary '
+                'files and/or directories at the start and at the end of '
+                'the test are not the same. Orphan entries: {}'.format(
+                    ','.join(final_temporary_items - self._temporary_items)
+                )
+            )
+        super().tearDown()