+++ mayan/apps/testing/dependencies.py
@@ -8,6 +8,7 @@ BinaryDependency(
     module=__name__, name='geckodriver',
     path=DEFAULT_FIREFOX_GECKODRIVER_PATH
 )
+
 PythonDependency(
     environment=environment_testing, module=__name__, name='coverage',
     version_string='==7.5.4'
@@ -23,8 +24,4 @@ PythonDependency(
 PythonDependency(
     environment=environment_testing, module=__name__, name='selenium',
     version_string='==3.141.0'
-)
-PythonDependency(
-    environment=environment_testing, module=__name__, name='psutil',
-    version_string='==6.0.0'
 )