+++ mayan/apps/testing/tests/base.py
@@ -15,6 +15,10 @@ from mayan.apps.events.tests.mixins.event_mixins import EventTestCaseMixin
 from mayan.apps.logging.tests.mixins import TestCaseMixinSilenceLogger
 from mayan.apps.permissions.tests.mixins import PermissionTestCaseMixin
 from mayan.apps.smart_settings.tests.mixins import SettingTestMixin
+from mayan.apps.storage.tests.mixins import (
+    DescriptorLeakCheckTestCaseMixin, OpenFileCheckTestCaseMixin,
+    TempfileCheckTestCasekMixin
+)
 from mayan.apps.views.tests.mixins import (
     ClientMethodsTestCaseMixin, DownloadTestCaseMixin, TestViewTestCaseMixin
 )
@@ -22,11 +26,7 @@ from mayan.apps.user_management.tests.mixins.user_mixins import UserTestMixin
 
 from ..literals import EXCLUDE_TEST_TAG
 
-from .mixins import (
-    ContentTypeCheckTestCaseMixin, DelayTestCaseMixin,
-    DescriptorLeakCheckTestCaseMixin, OpenFileCheckTestCaseMixin,
-    TempfileCheckTestCasekMixin
-)
+from .mixins import ContentTypeCheckTestCaseMixin, DelayTestCaseMixin
 
 
 class BaseTestCaseMixin(